+++ src/Routes.tsx
@@ -1,14 +1,19 @@
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
+import Counter from './pages/Counter';
 import Home from './pages/Home'
 
 
 export default function Routes() {
 
 	return (
-		<BrowserRouter>
-				<Switch> 
-					<Route path="/" exact component={Home} />
-				</Switch>
-		</BrowserRouter>
+
+		<>
+
+			<Switch>
+				<Route path="/counter"><Counter /></Route>
+				<Route path="/" ><Home /></Route>
+			</Switch>
+
+		</>
 	);
 }