+++ src/app/store.ts
@@ -1,0 +1,6 @@
+import { compose, createStore } from 'redux'
+import { rootReducer } from './reducers/reducers';
+
+// const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
+export default createStore(rootReducer)