+++ src/pages/Home.tsx
@@ -1,3 +1,12 @@
-const Home = () => <h1>HELLO FROM HOME!!!!!</h1>;
+import { Link } from "react-router-dom";
+
+const Home = () => {
+  return (
+    <>
+      <h1>HELLO FROM HOME!</h1>
+      <Link to="/counter">Click Me</Link>
+    </>
+  )
+};
 
 export default Home;